Create HttpClient as an interface.
Split the interface for HttpClient from the class,
making the UrlConnectionHttpClient the only impl.
